src/vision/scene_analyzer.py
"""
src/vision/scene_analyzer.py
Analyzes frames using computer vision models (BLIP, object detection)
"""

import logging
import numpy as np
from typing import Dict, List, Optional
from PIL import Image

logger = logging.getLogger(__name__)


class SceneAnalyzer:
    """Analyzes video frames for scene understanding"""
    
    def __init__(self, model_name: str = "blip-base"):
        """
        Initialize scene analyzer with vision model
        
        Args:
            model_name: Name of the vision model to use
        """
        self.model_name = model_name
        self.model = None
        self.processor = None
        
        self._load_model()
    
    def _load_model(self):
        """Load vision model (BLIP for scene captioning)"""
        try:
            from transformers import BlipProcessor, BlipForConditionalGeneration
            
            logger.info(f"Loading vision model: {self.model_name}")
            
            # Correct model paths for BLIP
            model_map = {
                "blip-base": "Salesforce/blip-image-captioning-base",
                "blip-large": "Salesforce/blip-image-captioning-large"
            }
            
            model_path = model_map.get(self.model_name, "Salesforce/blip-image-captioning-base")
            
            # Load BLIP model for image captioning
            self.processor = BlipProcessor.from_pretrained(model_path)
            self.model = BlipForConditionalGeneration.from_pretrained(model_path)
            
            # Move to GPU if available
            import torch
            if torch.cuda.is_available():
                self.model = self.model.to("cuda")
                logger.info("Model loaded on GPU")
            else:
                logger.info("Model loaded on CPU")
            
            logger.info("Vision model loaded successfully")
            
        except ImportError:
            logger.warning("transformers not installed, using fallback analyzer")
            self._use_fallback()
        except Exception as e:
            logger.error(f"Failed to load vision model: {e}")
            self._use_fallback()
    
    def _use_fallback(self):
        """Use simple fallback analyzer if model loading fails"""
        logger.info("Using fallback scene analyzer")
        self.model = None
        self.processor = None
    
    def analyze(self, frame: np.ndarray) -> Optional[Dict]:
        """
        Analyze a frame and return scene information
        
        Args:
            frame: Frame in BGR format from OpenCV
            
        Returns:
            Dictionary containing scene analysis results
        """
        try:
            if self.model is None:
                return self._fallback_analysis(frame)
            
            # Convert BGR to RGB
            rgb_frame = frame[:, :, ::-1]
            pil_image = Image.fromarray(rgb_frame)
            
            # Generate scene description
            description = self._generate_caption(pil_image)
            
            # Detect objects (placeholder for future enhancement)
            objects = self._detect_objects(frame)
            
            # Analyze activity (placeholder for future enhancement)
            activity = self._analyze_activity(frame)
            
            result = {
                'description': description,
                'objects': objects,
                'activity': activity,
                'confidence': 0.85  # Placeholder
            }
            
            logger.debug(f"Scene analysis: {description}")
            return result
            
        except Exception as e:
            logger.error(f"Error during scene analysis: {e}")
            return None
    
    def _generate_caption(self, image: Image.Image) -> str:
        """Generate natural language caption for image"""
        try:
            import torch
            
            inputs = self.processor(image, return_tensors="pt")
            
            # Move to same device as model
            if torch.cuda.is_available():
                inputs = {k: v.to("cuda") for k, v in inputs.items()}
            
            # Generate caption
            output = self.model.generate(**inputs, max_length=50)
            caption = self.processor.decode(output[0], skip_special_tokens=True)
            
            return caption
            
        except Exception as e:
            logger.error(f"Caption generation failed: {e}")
            return "Unable to describe scene"
    
    def _detect_objects(self, frame: np.ndarray) -> List[Dict]:
        """
        Detect objects in frame
        Placeholder for future object detection integration
        """
        # TODO: Integrate YOLO or similar for object detection
        return []
    
    def _analyze_activity(self, frame: np.ndarray) -> Optional[str]:
        """
        Analyze user activity from frame
        Placeholder for future activity recognition
        """
        # TODO: Integrate action recognition model
        return None
    
    def _fallback_analysis(self, frame: np.ndarray) -> Dict:
        """Simple fallback analysis when model not available"""
        # Basic image statistics
        brightness = np.mean(frame)
        
        description = "Person in frame" if brightness > 50 else "Low light scene"
        
        return {
            'description': description,
            'objects': [],
            'activity': None,
            'confidence': 0.3
        }